Feature/set map screen to actual data

+++ app/src/main/java/com/inveitix/android/clue/cmn/Door.java
@@ -20,16 +20,16 @@ public class Door {
         this.connectedTo = connectedTo;
     }
 
-    public double getX() {
-        return x;
+    public float getX() {
+        return (float) x;
     }
 
     public void setX(double x) {
         this.x = x;
     }
 
-    public double getY() {
-        return y;
+    public float getY() {
+        return (float) y;
     }
 
     public void setY(double y) {
